MemoryMonitor: move to proper location in Marlin
