+++ source/src/MemoryMonitor.cc
@@ -1,5 +1,5 @@
 /* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
-#include "MemoryMonitor.h"
+#include "marlin/MemoryMonitor.h"
 
 #include "sys/types.h"
